+++ packages/connect-wallet-modal/src/components/WalletsModalForEth/WalletsModalForEth.tsx
@@ -100,9 +100,7 @@ function getWalletsButtons(
   });
 }
 
-export function WalletsModalForEth(
-  props: WalletsModalForEthProps,
-): JSX.Element {
+export function WalletsModalForEth(props: WalletsModalForEthProps) {
   const { walletDataList } = useReefKnotContext();
 
   return (