+++ packages/web3-react/src/helpers/index.ts
@@ -3,3 +3,4 @@ export * from './ua';
 export * from './openWindow';
 export { default as isUrl } from './isUrl';
 export * from './interceptLedgerError';
+export * from './unsupportedChainError';