+++ packages/reef-knot/CHANGELOG.md
@@ -1,5 +1,15 @@
 # reef-knot
 
+## 1.10.6
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+  - @reef-knot/connect-wallet-modal@1.10.0
+  - @reef-knot/core-react@1.7.0
+  - @reef-knot/web3-react@1.8.0
+
 ## 1.10.5
 
 ### Patch Changes